# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sudoku Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /puzzles/new:
    get:
      operationId: generatePuzzle
      tags: [puzzles]
      summary: Generate a new Sudoku puzzle
      parameters:
        - name: difficulty
          in: query
          required: false
          schema:
            type: string
            enum: [easy, medium, hard, expert]
            default: medium
        - name: gridSize
          in: query
          required: false
          schema:
            type: integer
            enum: [3, 4, 9, 16]
            default: 9
      responses:
        "200":
          description: A new puzzle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Puzzle"
...
  /games:
    post:
      operationId: createGame
      tags: [games]
      summary: Start a new game session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Game started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a game session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
    patch:
      operationId: saveGame
      tags: [games]
      summary: Save game progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameProgress"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}/complete:
    post:
      operationId: completeGame
      tags: [games]
      summary: Mark a game as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameCompletion"
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
  /challenges:
    post:
      operationId: createChallenge
      tags: [challenges]
      summary: Create a new challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeInput"
      responses:
        "201":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeDetail"
  /challenges/for/{profileId}:
    get:
      operationId: getChallengesForProfile
      tags: [challenges]
      summary: Get all challenges for a profile (incoming and outgoing)
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChallengeDetail"
  /challenges/{id}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get a specific challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeDetail"
        "404":
          description: Not found
  /challenges/{id}/respond:
    patch:
      operationId: respondToChallenge
      tags: [challenges]
      summary: Accept or decline a challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeResponse"
      responses:
        "200":
          description: Updated challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Puzzle:
      type: object
      required: [id, difficulty, gridSize, grid, solution]
      properties:
        id:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard, expert]
        gridSize:
          type: integer
          enum: [3, 4, 9, 16]
        grid:
          type: string
          description: "gridSize*gridSize character string, 0 = empty; 16x16 uses hex chars (a-g for 10-16)"
        solution:
          type: string
        createdAt:
          type: string
    Game:
      type: object
      required: [id, puzzleId, status, currentGrid, elapsedSeconds, mistakeCount, createdAt]
      properties:
        id:
          type: integer
        profileId:
          type: ["integer", "null"]
        puzzleId:
          type: integer
        puzzle:
          $ref: "#/components/schemas/Puzzle"
        status:
          type: string
          enum: [active, completed, abandoned]
        currentGrid:
          type: string
        elapsedSeconds:
          type: integer
        mistakeCount:
          type: integer
        hintsUsed:
          type: integer
        points:
          type: ["integer", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    GameInput:
      type: object
      required: [puzzleId]
      properties:
        profileId:
          type: integer
        puzzleId:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard, expert]
    GameProgress:
      type: object
      properties:
        currentGrid:
          type: string
        elapsedSeconds:
          type: integer
        mistakeCount:
          type: integer
        hintsUsed:
          type: integer
    GameCompletion:
      type: object
      required: [elapsedSeconds]
      properties:
        elapsedSeconds:
          type: integer
        mistakeCount:
          type: integer
        hintsUsed:
          type: integer
...
    ChallengeInput:
      type: object
      required: [challengerId, challengedId, difficulty, gridSize]
      properties:
        challengerId:
          type: integer
        challengedId:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard, expert]
        gridSize:
          type: integer
          enum: [3, 4, 9, 16]
    ChallengeResponse:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, decline]
        profileId:
          type: integer
    ChallengeDetail:
      type: object
      required: [id, challengerId, challengedId, puzzleId, status, createdAt, challengerUsername, challengedUsername]
      properties:
        id:
          type: integer
        challengerId:
          type: integer
        challengedId:
          type: integer
        puzzleId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, declined, completed]
        challengerGameId:
          type: ["integer", "null"]
        challengedGameId:
          type: ["integer", "null"]
        winnerId:
          type: ["integer", "null"]
        challengerUsername:
          type: string
        challengedUsername:
          type: string
        challengerAvatar:
          type: ["string", "null"]
        challengedAvatar:
          type: ["string", "null"]
        challengerPoints:
          type: ["integer", "null"]
        challengedPoints:
          type: ["integer", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard, expert]
        gridSize:
          type: integer
          enum: [3, 4, 9, 16]
        createdAt:
          type: string