# lib/api-zod/src/generated/types/challengeDetail.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Sudoku Game API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeDetailDifficulty } from './challengeDetailDifficulty';
import type { ChallengeDetailGridSize } from './challengeDetailGridSize';
import type { ChallengeDetailStatus } from './challengeDetailStatus';

export interface ChallengeDetail {
  id: number;
  challengerId: number;
  challengedId: number;
  puzzleId: number;
  status: ChallengeDetailStatus;
  /** @nullable */
  challengerGameId?: number | null;
  /** @nullable */
  challengedGameId?: number | null;
  /** @nullable */
  winnerId?: number | null;
  challengerUsername: string;
  challengedUsername: string;
  /** @nullable */
  challengerAvatar?: string | null;
  /** @nullable */
  challengedAvatar?: string | null;
  /** @nullable */
  challengerPoints?: number | null;
  /** @nullable */
  challengedPoints?: number | null;
  difficulty?: ChallengeDetailDifficulty;
  gridSize?: ChallengeDetailGridSize;
  createdAt: string;
}