# lib/api-zod/src/generated/types/challengeInput.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Sudoku Game API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeInputDifficulty } from './challengeInputDifficulty';
import type { ChallengeInputGridSize } from './challengeInputGridSize';

export interface ChallengeInput {
  challengerId: number;
  challengedId: number;
  difficulty: ChallengeInputDifficulty;
  gridSize: ChallengeInputGridSize;
}