# lib/api-zod/src/generated/types/challengeResponse.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Sudoku Game API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeResponseAction } from './challengeResponseAction';

export interface ChallengeResponse {
  action: ChallengeResponseAction;
  profileId?: number;
}